services/insights_service/requirements.txt
# Insights Service Requirements

# FastAPI and ASGI server
uvicorn==0.35.0

# Google Gemini SDK
google-generativeai>=0.3.0

# Metrics and monitoring
prometheus-client==0.19.0

# Test dependencies
httpx>=0.25.0
